wip: align code block style
