+++ src/lib/styles/index.css
@@ -1,3 +1,4 @@
 @import "./tokens.css";
 @import "./base.css";
 @import "./vp-doc.css";
+@import "./vp-code.css";
