+++ src/lib/styles/vp-code.css
@@ -1,0 +1,186 @@
+/* VitePress Code Block Styles */
+
+/* Code block variables */
+:root {
+	--vp-code-line-height: 1.7;
+	--vp-code-font-size: 0.875em;
+	--vp-code-color: var(--vp-c-brand-1);
+	--vp-code-link-color: var(--vp-c-brand-1);
+	--vp-code-c-hover-bg: var(--vp-c-bg-soft);
+	--vp-code-tab-divider: var(--vp-c-divider);
+	--vp-code-copy-code-border-color: var(--vp-c-divider);
+	--vp-code-copy-code-active-text: var(--vp-c-text-2);
+	--vp-code-copy-code-hover-border-color: var(--vp-c-divider);
+	--vp-code-copy-code-hover-bg: var(--vp-c-bg-soft);
+	--vp-code-copy-code-active-bg: var(--vp-c-bg-soft);
+	--vp-code-line-highlight-color: var(--vp-c-default-soft);
+	--vp-code-line-number-color: var(--vp-c-text-3);
+	--vp-code-line-diff-add-color: var(--vp-c-success-soft);
+	--vp-code-line-diff-add-symbol-color: var(--vp-c-success-1);
+	--vp-code-line-diff-remove-color: var(--vp-c-danger-soft);
+	--vp-code-line-diff-remove-symbol-color: var(--vp-c-danger-1);
+	--vp-code-line-warning-color: var(--vp-c-warning-soft);
+	--vp-code-line-error-color: var(--vp-c-danger-soft);
+	--vp-code-copy-opacity: 0;
+	--vp-code-copy-opacity-hover: 1;
+	--vp-code-lang-color: var(--vp-c-text-3);
+}
+
+/* Main code block container */
+.vp-doc div[class*="language-"] {
+	position: relative;
+	margin: 16px -24px;
+	background-color: var(--vp-code-block-bg);
+	overflow-x: auto;
+	transition: background-color 0.5s;
+}
+
+@media (min-width: 640px) {
+	.vp-doc div[class*="language-"] {
+		border-radius: 8px;
+		margin: 16px 0;
+	}
+}
+
+/* Pre element */
+.vp-doc [class*="language-"] pre {
+	position: relative;
+	z-index: 1;
+	margin: 0;
+	padding: 20px 0;
+	background: transparent;
+	overflow-x: auto;
+	text-align: left;
+}
+
+/* Code element */
+.vp-doc [class*="language-"] code {
+	display: block;
+	padding: 0 24px;
+	width: fit-content;
+	min-width: 100%;
+	line-height: var(--vp-code-line-height);
+	font-size: var(--vp-code-font-size);
+	color: var(--vp-code-block-color);
+	transition: color 0.5s;
+}
+
+/* Language label */
+.vp-doc div[class*="language-"] > span.lang {
+	position: absolute;
+	top: 2px;
+	right: 8px;
+	z-index: 2;
+	font-size: 12px;
+	font-weight: 500;
+	color: var(--vp-code-lang-color);
+	transition:
+		color 0.4s,
+		opacity 0.4s;
+}
+
+/* Copy button */
+.vp-doc div[class*="language-"] > button.copy {
+	direction: ltr;
+	position: absolute;
+	top: 12px;
+	right: 12px;
+	z-index: 3;
+	border: 1px solid var(--vp-code-copy-code-border-color);
+	border-radius: 4px;
+	width: 40px;
+	height: 40px;
+	background-color: var(--vp-code-block-bg);
+	opacity: var(--vp-code-copy-opacity);
+	cursor: pointer;
+	background-image: var(--vp-icon-copy);
+	background-position: 50%;
+	background-size: 20px;
+	background-repeat: no-repeat;
+	transition:
+		border-color 0.25s,
+		background-color 0.25s,
+		opacity 0.25s;
+}
+
+.vp-doc div[class*="language-"]:hover > button.copy,
+.vp-doc div[class*="language-"] > button.copy:focus {
+	opacity: var(--vp-code-copy-opacity-hover);
+}
+
+.vp-doc div[class*="language-"] > button.copy:hover,
+.vp-doc div[class*="language-"] > button.copy.copied {
+	border-color: var(--vp-code-copy-code-hover-border-color);
+	background-color: var(--vp-code-copy-code-hover-bg);
+}
+
+.vp-doc div[class*="language-"] > button.copy.copied,
+.vp-doc div[class*="language-"] > button.copy:hover.copied {
+	border-radius: 0 4px 4px 0;
+	background-color: var(--vp-code-copy-code-active-bg);
+	background-image: var(--vp-icon-copied);
+}
+
+.vp-doc div[class*="language-"] > button.copy.copied::before,
+.vp-doc div[class*="language-"] > button.copy:hover.copied::before {
+	position: relative;
+	top: -1px;
+	transform: translateX(-65px);
+	display: flex;
+	align-items: center;
+	border: 1px solid var(--vp-code-copy-code-hover-border-color);
+	border-right: 0;
+	border-radius: 4px 0 0 4px;
+	padding: 0 10px;
+	width: fit-content;
+	height: 40px;
+	text-align: center;
+	font-size: 12px;
+	font-weight: 500;
+	color: var(--vp-code-copy-code-active-text);
+	background-color: var(--vp-code-copy-code-active-bg);
+	white-space: nowrap;
+	content: "Copied";
+}
+
+/* Icons using CSS (since we can't easily use SVGs in CSS custom properties) */
+:root {
+	--vp-icon-copy: url("data:image/svg+xml,%3csvg width='16' height='16' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3e%3cpath d='M5 22C4.45 22 3.97917 21.8042 3.5875 21.4125C3.19583 21.0208 3 20.55 3 20V6H5V20H16V22H5ZM9 18C8.45 18 7.97917 17.8042 7.5875 17.4125C7.19583 17.0208 7 16.55 7 16V4C7 3.45 7.19583 2.97917 7.5875 2.5875C7.97917 2.19583 8.45 2 9 2H18C18.55 2 19.0208 2.19583 19.4125 2.5875C19.8042 2.97917 20 3.45 20 4V16C20 16.55 19.8042 17.0208 19.4125 17.4125C19.0208 17.8042 18.55 18 18 18H9ZM9 16H18V4H9V16Z' fill='%23999'/%3e%3c/svg%3e");
+	--vp-icon-copied: url("data:image/svg+xml,%3csvg width='16' height='16' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'%3e%3cpath d='M9.55 18L3.85 12.3L5.275 10.875L9.55 15.15L18.725 5.975L20.15 7.4L9.55 18Z' fill='%2300b894'/%3e%3c/svg%3e");
+}
+
+/* Dark mode theme switching */
+.dark .shiki span {
+	color: var(--shiki-dark, inherit);
+}
+
+html:not(.dark) .shiki span {
+	color: var(--shiki-light, inherit);
+}
+
+/* VitePress adaptive theme support */
+.vp-adaptive-theme .shiki {
+	background-color: transparent !important;
+}
+
+.vp-adaptive-theme .shiki span {
+	color: var(--shiki-light);
+}
+
+.dark .vp-adaptive-theme .shiki span {
+	color: var(--shiki-dark);
+}
+
+/* Code line specific styling */
+.vp-doc [class*="language-"] .line {
+	/* Additional line-specific styles if needed */
+}
+
+/* Highlighted lines */
+.vp-doc [class*="language-"] code .highlighted {
+	background-color: var(--vp-code-line-highlight-color);
+	margin: 0 -24px;
+	padding: 0 24px;
+	width: calc(100% + 2 * 24px);
+	display: inline-block;
+}
